+++ web/src/test/handlers.ts
@@ -8,6 +8,7 @@ import {
   makeBackup,
   makeBannedPlayer,
   makeCatalog,
+  makeClusterInfo,
   makeClusterStats,
   makeClusterView,
   makeConfig,
@@ -94,9 +95,7 @@ export const handlers = [
   }),
 
   http.get("/cluster", () => HttpResponse.json(makeClusterView())),
-  http.get("/cluster/info", () =>
-    HttpResponse.json({ clusterName: "homelab", version: "v1.31.0" }),
-  ),
+  http.get("/cluster/info", () => HttpResponse.json(makeClusterInfo())),
   http.get("/cluster/stats", () => HttpResponse.json(makeClusterStats())),
 
   // Servers